pyproject.toml
[tool.poetry]
name = "ought-ice-demo"
version = "0.1.0"
description = "Ought ICE library demo"
authors = ["Your Name <you@example.com>"]
license = "MIT"
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "~3.11"
ought-ice = "^0.5.0"
pydantic = "<2"
ipykernel = "^6.29.5"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
